+++ src/apis/problemStatementApi.js
@@ -2,4 +2,4 @@ import axios from 'axios';
 
 import { SERVER_URL } from 'constants/appConstants';
 
-export const getStatement = () => axios.get(`${SERVER_URL}/statementsList`);
+export const getStatement = () => axios.get(`${SERVER_URL}/statementsLis`);
