+++ src/core-components/index.js
@@ -1,4 +1,4 @@
-import Button from './Button';
-import Container from './Container';
+import Button from 'core-components/Button';
+import Container from 'core-components/Container';
 
 export { Button, Container };
